Week4_lesson_Realtime-Voice-Chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai-whisper
groq
gTTS

